+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fix Content-Disposition for error responses (#82)
 * Add pid to /cachestatus (#83)
 * Check CDB_QueryTable response before saving into cache (#83)
+* Use an expiring LRU cache for query tables (#83)
 
 1.3.4 (21/01/13)
 -----